Restructure components folder to group files into logical folders
