+++ frontend/src/app/components/features/calendarView/CalendarView.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.module.scss";
+@import "../../../variables.module.scss";
 
 // --- Header Part -----------------
 .headerToolbar {
