+++ frontend/src/app/components/features/classDetail/ClassDetail.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.module.scss";
+@import "../../../variables.module.scss";
 
 .classCard {
   padding: 0 25px 25px 25px;
