+++ frontend/src/app/components/features/classesTable/ClassesTable.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.module.scss";
+@import "../../../variables.module.scss";
 
 .classesTable {
   &__wrapper {
