+++ frontend/src/app/components/features/classesTable/ClassesTable.tsx
@@ -3,8 +3,8 @@ import {
   isPastClassDateTime,
   formatShortDate,
   formatTime24Hour,
-} from "../helper/dateUtils";
-import ActionButton from "./elements/buttons/actionButton/ActionButton";
+} from "../../../helper/dateUtils";
+import ActionButton from "../../elements/buttons/actionButton/ActionButton";
 import styles from "./ClassesTable.module.scss";
 import Image from "next/image";
 import { InformationCircleIcon } from "@heroicons/react/24/solid";
