+++ frontend/src/app/customers/[id]/classes/[classId]/page.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState, useCallback } from "react";
 import { cancelClass, getClassesByCustomerId } from "@/app/helper/classesApi";
-import ClassDetail from "@/app/components/ClassDetail";
-import ClassesTable from "@/app/components/ClassesTable";
+import ClassDetail from "@/app/components/features/classDetail/ClassDetail";
+import ClassesTable from "@/app/components/features/classesTable/ClassesTable";
 import RedirectButton from "@/app/components/elements/buttons/redirectButton/RedirectButton";
 import {
   isPastClassDateTime,
